+++ ui/v2.5/src/components/Wall/WallPanel.tsx
@@ -1,10 +1,7 @@
-import React, { MouseEvent, useCallback, useContext, useEffect, useMemo, useState } from "react";
+import React, { MouseEvent } from "react";
 import * as GQL from "src/core/generated-graphql";
 import { SceneQueue } from "src/models/sceneQueue";
 import { WallItem, WallItemData, WallItemType } from "./WallItem";
-import Gallery, { PhotoProps, renderImageClickHandler, RenderImageProps } from "react-photo-gallery";
-import { ConfigurationContext } from "src/hooks/Config";
-import { objectTitle } from "src/core/files";
 
 interface IWallPanelProps<T extends WallItemType> {
   type: T;
@@ -76,192 +73,5 @@ export const MarkerWallPanel: React.FC<IMarkerWallPanelProps> = ({
 }) => {
   return (
     <WallPanel type="sceneMarker" data={markers} clickHandler={clickHandler} />
-  );
-};
-
-interface ISceneWallPanelProps {
-  scenes: GQL.SlimSceneDataFragment[];
-  sceneQueue?: SceneQueue;
-  clickHandler?: (e: MouseEvent, item: GQL.SlimSceneDataFragment) => void;
-}
-
-export const SceneWallPanel: React.FC<ISceneWallPanelProps> = ({
-  scenes,
-  sceneQueue,
-  clickHandler,
-}) => {
-  return (
-    // <WallPanel
-    //   type="scene"
-    //   data={scenes}
-    //   sceneQueue={sceneQueue}
-    //   clickHandler={clickHandler}
-    // />
-    <SceneWall
-      scenes={scenes}
-      handleImageOpen={() => {}}
-    />
-  );
-};
-
-interface IBackupSrc {
-  backupSrc?: string;
-  onError?: (photo: PhotoProps<IBackupSrc>) => void;
-}
-
-interface ISceneWallItemProps {
-  margin?: string;
-  index: number;
-  photo: PhotoProps<IBackupSrc>;
-  onClick: renderImageClickHandler | null;
-  direction: "row" | "column";
-  top?: number;
-  left?: number;
-}
-
-export const SceneWallItem: React.FC<RenderImageProps> = (
-  props: ISceneWallItemProps
-) => {
-  type style = Record<string, string | number | undefined>;
-  var imgStyle: style = {
-    margin: props.margin,
-    display: "block",
-  };
-
-  if (props.direction === "column") {
-    imgStyle.position = "absolute";
-    imgStyle.left = props.left;
-    imgStyle.top = props.top;
-  }
-
-  var handleClick = function handleClick(
-    event: React.MouseEvent<Element, MouseEvent>
-  ) {
-    if (props.onClick) {
-      props.onClick(event, { index: props.index });
-    }
-  };
-
-  const video = props.photo.src.includes("preview");
-  const VideoPreview = "video";
-  const ImagePreview = video ? "video" : "img";
-
-  const previewProps = {
-    loop: video,
-    muted: video,
-    autoPlay: video,
-    key: props.photo.key,
-    style: imgStyle,
-    src: props.photo.src,
-    width: props.photo.width,
-    height: props.photo.height,
-    alt: props.photo.alt,
-    onClick: handleClick,
-  };
-
-  // if (isMissing) {
-  //   // show the image if the video preview is unavailable
-  //   if (props.photo.backupSrc) {
-  //     return <img {...previewProps} src={props.photo.backupSrc} />;
-  //   }
-
-  //   return (
-  //     <div className="wall-item-media wall-item-missing">
-  //       Pending preview generation
-  //     </div>
-  //   );
-  // }
-
-  return (
-    <ImagePreview
-      loop={video}
-      muted={video}
-      autoPlay={video}
-      key={props.photo.key}
-      style={imgStyle}
-      src={props.photo.src}
-      width={props.photo.width}
-      height={props.photo.height}
-      alt={props.photo.alt}
-      onClick={handleClick}
-      onError={() => {
-        props.photo.onError?.(props.photo);
-      }}
-    />
-  );
-};
-
-interface ISceneWallProps {
-  scenes: GQL.SlimSceneDataFragment[];
-  handleImageOpen: (index: number) => void;
-}
-
-const SceneWall: React.FC<ISceneWallProps> = ({ scenes, handleImageOpen }) => {
-  const { configuration } = useContext(ConfigurationContext);
-  const uiConfig = configuration?.ui;
-
-  const [erroredImgs, setErroredImgs] = useState<string[]>([]);
-
-  const handleError = useCallback(
-    (photo: PhotoProps<IBackupSrc>) => {
-      setErroredImgs((prev) => [...prev, photo.src]);
-    },
-    []
-  );
-
-  useEffect(() => {
-    setErroredImgs([]);
-  }, [scenes]);
-
-  const photos: PhotoProps<IBackupSrc>[] = useMemo(() => {
-    return scenes.filter((s) => {
-      return s.files.length > 0 && s.files[0].width > 0 && s.files[0].height > 0;
-    }).map((s, index) => {
-      const width = s.files[0].width;
-      const height = s.files[0].height;
-
-      return {
-        src:
-          s.paths.preview && !erroredImgs.includes(s.paths.preview)
-            ? s.paths.preview!
-            : s.paths.screenshot!,
-        width,
-        height,
-        tabIndex: index,
-        key: s.id,
-        loading: "lazy",
-        className: "gallery-image",
-        alt: objectTitle(s),
-        onError: handleError,
-      };
-    });
-  }, [scenes, erroredImgs, handleError]);
-
-  const showLightboxOnClick = useCallback(
-    (event, { index }) => {
-      // handleImageOpen(index);
-    },
-    [handleImageOpen]
-  );
-
-  function columns(containerWidth: number) {
-    let preferredSize = 300;
-    let columnCount = containerWidth / preferredSize;
-    return Math.round(columnCount);
-  }
-
-  return (
-    <div className="gallery">
-      {photos.length ? (
-        <Gallery
-          photos={photos}
-          renderImage={SceneWallItem}
-          onClick={showLightboxOnClick}
-          margin={uiConfig?.imageWallOptions?.margin!}
-          direction={uiConfig?.imageWallOptions?.direction!}
-          columns={columns}
-        />
-      ) : null}
-    </div>
   );
 };
